api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
httpx==0.27.0
python-dotenv==1.0.1
supabase==2.9.1
sajupy==0.3.0
openai>=1.30.0
